Get rid of very old bug with GPS or battery driver fail to init on boot.
There is no fix for that for now, but in very rare conditions on REBOOT
and USB plugged, GPS and Battery driver fail to init on boot.

and blue load line bar is stuck at max for 3~5 sec and then boot.

But GPS will not work.

So check for that, and if detected bug, REBOOT to fix it.