HACK, for now to fix USB transfer to External SD, use deadline gov.
After many tests i have found that most stable I/O got for large file transfer
is DEADLINE. but it's needed only on PC to USB > EXT SD transfers, then files
transfered and saved, and USB driver do not fail.

BFQ and others, working OK for transfers to Internal SD.

So this is good solution for now.

Also set this in RECOVERY.